rename sectionID to SectionSelectors. refactor addFilms method
