+++ .idea/workspace.xml
@@ -5,9 +5,16 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="154ebf74-2b4c-4c71-a8b2-acfb77888ecd" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/core/components/filmCardComponent.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/components/filmCardComponent.ts" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/core/components/searchLineComponent.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/components/searchLineComponent.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/inspectionProfiles/Project_Default.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/vcs.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/controller/controller.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/controller/controller.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/core/constants/SectionID.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/constants/SectionSelectors.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/core/router.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/router.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/style/global/global.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/style/global/global.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/views/favoritesView.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/views/favoritesView.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/views/filmsListView.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/views/filmsListView.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/views/responseErrorView.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/views/responseErrorView.ts" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -134,7 +141,9 @@
       <workItem from="1639041758032" duration="414000" />
       <workItem from="1639050403992" duration="4005000" />
       <workItem from="1639137544434" duration="14247000" />
-      <workItem from="1639375232987" duration="5859000" />
+      <workItem from="1639375232987" duration="7135000" />
+      <workItem from="1639400244527" duration="81000" />
+      <workItem from="1639400338552" duration="595000" />
     </task>
     <servers />
   </component>
