+++ src/style/global/global.scss
@@ -74,7 +74,7 @@ html, body, #root {
   justify-content: space-around;
   align-items: center;
 
-  height: $header-height;
+  height: $footer-height;
 
   font-size: 15px;
   background: $background-dark;
